chore:    js/utility.js -> scripts/utils.js 변경 	renamed:    js/utility.js -> scripts/utils.js 	modified:   signin/index.js 	modified:   signup/index.js
